requirements.txt
absl-py==1.2.0
astor==0.8.1
async-timeout==4.0.2
cached-property==1.5.2
certifi==2021.5.30
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
Deprecated==1.2.13
Django==2.2.5
django-cors-headers==3.10.1
django-redis==5.2.0
djangorestframework==3.13.1
drf-yasg2==1.19.4
gast==0.5.3
grpcio==1.47.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
jieba==0.42.1
Jinja2==3.0.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.7
MarkupSafe==2.0.1
numpy==1.16.4
packaging==21.3
pandas==1.1.5
protobuf==4.21.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
redis==4.3.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.0.3
wincertstore==0.2
wrapt==1.14.1
zipp==3.6.0
